+++ lib/go-tc/cachegroups.go
@@ -95,3 +95,4 @@ type CachegroupQueueUpdatesRequest struct {
 	CDN    *CDNName         `json:"cdn"`
 	CDNID  *util.JSONIntStr `json:"cdnId"`
 }
+